% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pick_lock.R
\name{pick_lock}
\alias{pick_lock}
\title{Solve a lockpicking puzzle}
\usage{
pick_lock(start, ops, goal = rep(0, length(start)), max_abs = 3, refine = TRUE)
}
\arguments{
\item{start}{The initial state of the lock}

\item{ops}{The effects of moving each lock shackle}

\item{goal}{The desired end state of the lock (defaults to all zeros)}

\item{max_abs}{The maximum allowed pin location in a shackle (defaults to 3)}

\item{refine}{Should the solver try to simplify the result? (default \code{TRUE})}
}
\description{
Find a shortest sequence of moves from \code{start} to \code{goal}. Returns a list with
\code{states} (start ... goal) and \code{moves} (data.frame of op + sign). Errors if no
path exists.
}
