% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_solution.R
\name{print.solution}
\alias{print.solution}
\title{Print a lockpicking solution}
\usage{
\method{print}{solution}(x, ...)
}
\arguments{
\item{x}{A \code{solution}, as returned by \code{\link[=pick_lock]{pick_lock()}}.}

\item{...}{Ignored.}
}
\description{
Print a lockpicking solution
}
